Some house keeping (renamed mode for core options, etc)
